pyproject.toml
[project]
name = "phealthassistant"
version = "0.1.0"
description = "Healthcare RAG AI Agent"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-generativeai>=0.8.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["src"]
